+++ changelog.md
@@ -3,6 +3,7 @@
 ## Upcoming
 ### Changes
 * Experience to level up reduced (50 -> 40)
+* Insomnia game rule is now set to false when tiredness is enabled
 
 ## 4.11.4
 ### Changes